separate module for approval checking utilities
